Add extra conditions for final loop
A relatively new feature is to detect and delay the final notification being sent to give frigate enough time to save the final snapshot. This is mostly relevant when people use features in frigate like boundary box and timestamp as there can be a delay in processing them. So a 10 second delay was implemented to the final loop in the notification.

This change adds two extra conditions. The delay to the final loop will only be implemented if the user has attachments set to snapshot and update thumbnail set to tru, otherwise it is affecting users who won't get a notification anyway.