Add TTS text helper option to prevent multi TTS
When an Event triggers a TTS notification, the Event ID is prepended
onto this input text helper. On subsequent updates to the event, this
helper is first checked to see if it contains the event ID, indicating
that a TTS alert has already fired. If this is the case, TTS will not
fire again for this event.