Migrate legacy File notify service for domain
Implemented the following...

The File notify service(s) are migrated. A new notify entity is available now to replace each legacy notify service.

Update any automations to use the new notify.send_message service exposed with this new entity. When this is done, fix this issue and restart Home Assistant.

There is no depreciation notice yet, so there is time to test.